requirements.txt
matplotlib==3.7.2
numpy==1.23.5
Pillow==11.2.1
scikit_learn==1.7.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
